requirements.txt
tensorflow==2.3.0
tensorflow_text
tensorflow_hub
transformers
sklearn
absl-py
sacrebleu==1.3.1
pandas
more_itertools
numpy